1-1 Relationship

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Dotnet_Core_Web_API.Data;
 using Dotnet_Core_Web_API.Dto.CharacterDto;
 using Dotnet_Core_Web_API.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace Dotnet_Core_Web_API.Services.CharacterService
@@ -14,18 +16,24 @@ namespace Dotnet_Core_Web_API.Services.CharacterService
     {
         private readonly AppDbContext _context;
         private readonly IMapper _mapper;
-        public CharacterService(AppDbContext context, IMapper mapper)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public CharacterService(AppDbContext context, IMapper mapper, IHttpContextAccessor httpContextAccessor)
         {
+            this._httpContextAccessor = httpContextAccessor;
             this._mapper = mapper;
             this._context = context;
         }
 
+        private int GetUserId() => int.Parse(this._httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto addCharacter)
         {
             ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();
-            this._context.Characters.Add(this._mapper.Map<Character>(addCharacter));
+            Character character = this._mapper.Map<Character>(addCharacter);
+            character.User = await this._context.Users.FirstOrDefaultAsync(u => u.Id == this.GetUserId());
+            this._context.Characters.Add(character);
             this._context.SaveChanges();
-            List<Character> characters = await this._context.Characters.ToListAsync();
+            List<Character> characters = await this._context.Characters.Where(c => c.User.Id == this.GetUserId()).ToListAsync();
             response.Data = characters.Select(c => this._mapper.Map<GetCharacterDto>(c)).ToList();
             return response;
         }
@@ -35,11 +43,20 @@ namespace Dotnet_Core_Web_API.Services.CharacterService
             ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await this._context.Characters.FirstAsync(c => c.Id == id);
-                this._context.Characters.Remove(character);
-                this._context.SaveChanges();
-                List<Character> characters = await this._context.Characters.ToListAsync();
-                response.Data = characters.Select(c => this._mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await this._context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == this.GetUserId());
+                if (character != null)
+                {
+                    this._context.Characters.Remove(character);
+                    this._context.SaveChanges();
+
+                    List<Character> characters = await this._context.Characters.Where(c => c.User.Id == this.GetUserId()).ToListAsync();
+                    response.Data = characters.Select(c => this._mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
             }
             catch (Exception e)
             {
@@ -49,10 +66,10 @@ namespace Dotnet_Core_Web_API.Services.CharacterService
             return response;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> characters = await this._context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            List<Character> characters = await this._context.Characters.Where(c => c.User.Id == this.GetUserId()).ToListAsync();
             response.Data = characters.Select(c => this._mapper.Map<GetCharacterDto>(c)).ToList();
             return response;
         }
@@ -60,16 +77,8 @@ namespace Dotnet_Core_Web_API.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> response = new ServiceResponse<GetCharacterDto>();
-            try
-            {
-                Character character = await this._context.Characters.FirstAsync(c => c.Id == id);
-                response.Data = this._mapper.Map<GetCharacterDto>(character);
-            }
-            catch (Exception e)
-            {
-                response.Success = false;
-                response.Message = e.Message;
-            }
+            Character character = await this._context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == this.GetUserId());
+            response.Data = this._mapper.Map<GetCharacterDto>(character);
             return response;
         }
 
@@ -78,16 +87,24 @@ namespace Dotnet_Core_Web_API.Services.CharacterService
             ServiceResponse<GetCharacterDto> response = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await this._context.Characters.FirstAsync(c => c.Id == updateCharacter.Id);
-                character.Name = updateCharacter.Name;
-                character.HitPoints = updateCharacter.HitPoints;
-                character.Defense = updateCharacter.Defense;
-                character.Strength = updateCharacter.Strength;
-                character.Intelligence = updateCharacter.Intelligence;
-                character.Class = updateCharacter.Class;
-                this._context.Characters.Update(character);
-                this._context.SaveChanges();
-                response.Data = this._mapper.Map<GetCharacterDto>(character);
+                Character character = await this._context.Characters.Include(c => c.User).FirstAsync(c => c.Id == updateCharacter.Id);
+                if (character.User.Id == this.GetUserId())
+                {
+                    character.Name = updateCharacter.Name;
+                    character.HitPoints = updateCharacter.HitPoints;
+                    character.Defense = updateCharacter.Defense;
+                    character.Strength = updateCharacter.Strength;
+                    character.Intelligence = updateCharacter.Intelligence;
+                    character.Class = updateCharacter.Class;
+                    this._context.Characters.Update(character);
+                    await this._context.SaveChangesAsync();
+                    response.Data = this._mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
             }
             catch (Exception e)
             {
